refactor(GiniHealthSDK): Move BankTableCell class in a separate file.
PIA-2040
